requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.2.0
    # via httpx
attrs==23.2.0
    # via jsonschema
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charmed-kubeflow-chisme==0.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
exceptiongroup==1.2.0
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via lightkube
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.1.1
    # via jsonschema
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jsonschema==4.17.3
    # via serialized-data-interface
lightkube==0.15.0
    # via charmed-kubeflow-chisme
lightkube-models==1.29.0.6
    # via lightkube
markupsafe==2.1.3
    # via jinja2
ops==2.9.0
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pkgutil-resolve-name==1.3.10
    # via jsonschema
pyrsistent==0.20.0
    # via jsonschema
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via serialized-data-interface
ruamel-yaml==0.18.5
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
serialized-data-interface==0.7.0
    # via charmed-kubeflow-chisme
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
tenacity==8.2.3
    # via charmed-kubeflow-chisme
typing-extensions==4.9.0
    # via anyio
urllib3==2.1.0
    # via requests
websocket-client==1.7.0
    # via ops
zipp==3.17.0
    # via importlib-resources
